services/ingestion-platform/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic>=2.7.0
pydantic-settings>=2.3.0
motor==3.5.0
chromadb==0.5.3
httpx==0.27.0
anthropic>=0.40.0
mcp>=1.0.0
playwright>=1.45.0
beautifulsoup4>=4.12.0
pyyaml>=6.0.1
jsonschema>=4.22.0
croniter>=3.0.3
rank-bm25==0.2.2
numpy<2.0
pytest==8.2.2
pytest-asyncio==0.23.7
